Implement support for non-static member functions as tasks
All the functions now have new overload that takes pointer to an object
and pointer to a non-static member function that will be executed on the
object.

This allows using KAsync for example in constructor of a class and then
implementing each task as a member method, so that there is access to
some sort of context, or state of the object.