Introduce Job::error(ErrorHandler) to pass general error handler
The error() method only accepts a single argument, the error handler.
The error handler is only invoked if an error occured anywhere in the
execution chain and was not handled by any other error handler. If
there was no error (or the error was already handled), it will only
pass through the result of the previous task.

This is equivalent to calling

.then<PrevOut, PrevOut>(
    [](const PrevOut &in) -> PrevOut
    {
        return in;
    },
    errorHandler);

(it is also exactly what the method does internally ;-)