Avoid the member function overload with non-class types.
Summary:
In the akonadinext codebase we face the issue that either a member function
overload get's selected that then fails, if the return type is not a class
type. This workaround simply disables those overloads in case of non class
types.

Unfortunately I failed to write a test to reproduce the issue.
It may be that the issue is also better solved in detail::syncFuncHelper
and alike.

Test Plan: Tests pass and akonadinext builds again.

Reviewers: dvratil

Subscribers: