Added implicit conversion to Job<void>
This is necessary to be able to return a Job<void> after i.e. having a
.then<void, int>(), because in the return value we really don't care
about the int argument.
It's a bit of a hack necessary because the job assembly doesn't have
the type Job<Input, Output> with Input being from the initial
job, instead of the last continuation (.then clause).