Don't crash with users that have QT_NO_DEBUG disabled.
Since part of the library is in header files and parts not, we can't
make assumptions that we have the same state everywhere. (It used to
crash with the wrong combinations). With this patch we ensure that we
don't incur the overhead with QT_NO_DEBUG enabled, but otherwise let
the client/user of the library control the value.